Split maketrees out into a separate tool
